7553 data model event changes
