+++ Core/src/org/sleuthkit/autopsy/datamodel/DataSources.java
@@ -1,7 +1,7 @@
 /*
  * Autopsy Forensic Browser
  * 
- * Copyright 2011 Basis Technology Corp.
+ * Copyright 2011-2021 Basis Technology Corp.
  * Contact: carrier <at> sleuthkit <dot> org
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,7 +19,7 @@
 package org.sleuthkit.autopsy.datamodel;
 
 /**
- * Root node to store the data sources in a case
+ * An "Autopsy visitable item" that supplies a 
  */
 public class DataSources implements AutopsyVisitableItem {
