+++ Core/src/org/sleuthkit/autopsy/datamodel/DisplayableItemNodeVisitor.java
@@ -42,7 +42,7 @@ public interface DisplayableItemNodeVisitor<T> {
     /*
      * Data Sources Area
      */
-    T visit(DataSourcesNode in);
+    T visit(DataSourceFilesNode in);
 
     T visit(LayoutFileNode lfn);
 
@@ -201,7 +201,7 @@ public interface DisplayableItemNodeVisitor<T> {
 
     T visit(HostNode node);
 
-    T visit(DataSourcesByTypeNode node);
+    T visit(DataSourcesNode node);
     
     /*
      * Unsupported node
@@ -416,7 +416,7 @@ public interface DisplayableItemNodeVisitor<T> {
         }
 
         @Override
-        public T visit(DataSourcesNode in) {
+        public T visit(DataSourceFilesNode in) {
             return defaultVisit(in);
         }
 
@@ -571,7 +571,7 @@ public interface DisplayableItemNodeVisitor<T> {
         }
 
         @Override
-        public T visit(DataSourcesByTypeNode node) {
+        public T visit(DataSourcesNode node) {
             return defaultVisit(node);
         }
