+++ Core/src/org/sleuthkit/autopsy/datamodel/RootContentChildren.java
@@ -135,7 +135,7 @@ public class RootContentChildren extends Children.Keys<Object> {
 
         @Override
         public AbstractNode visit(DataSources i) {
-            return new DataSourcesNode(i.filteringDataSourceObjId());
+            return new DataSourceFilesNode(i.filteringDataSourceObjId());
         }
 
         @Override
@@ -202,7 +202,7 @@ public class RootContentChildren extends Children.Keys<Object> {
 
         @Override
         public AbstractNode visit(DataSourcesByType dataSourceHosts) {
-            return new DataSourcesByTypeNode();
+            return new DataSourcesNode();
         }
     }
 }
